Added Json ANNNOTAITON TO HANDLE cyclic output from many to many

+++ basic-concpets/conferenceSchedulingApp/src/main/java/com/conferenceschedulingapp/repositories/SessionRepo.java
@@ -3,6 +3,6 @@ package com.conferenceschedulingapp.repositories;
 import com.conferenceschedulingapp.models.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SessionRepo extends JpaRepository <Session, Long> {
+public interface SessionRepo extends JpaRepository <Session, Integer> {
 
 }
