+++ basic-concpets/conferenceSchedulingApp/src/main/java/com/conferenceschedulingapp/repositories/SpeakerRepo.java
@@ -3,5 +3,5 @@ package com.conferenceschedulingapp.repositories;
 import com.conferenceschedulingapp.models.Speaker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpeakerRepo extends JpaRepository <Speaker, Long> {
+public interface SpeakerRepo extends JpaRepository <Speaker, Integer> {
 }
